+++ judge-sdk/src/main/java/com/dwarfeng/judge/sdk/util/ServiceExceptionCodes.java
@@ -72,7 +72,12 @@ public final class ServiceExceptionCodes {
             new ServiceException.Code(offset(220), "invalid judgement modal update value");
     public static final ServiceException.Code INVALID_ALARM_MODAL_UPDATE_HAPPENED_DATE =
             new ServiceException.Code(offset(230), "invalid alarm modal update happened date");
-
+    public static final ServiceException.Code RECEIVER_FAILED =
+            new ServiceException.Code(offset(240), "receiver failed");
+    public static final ServiceException.Code RECEIVER_NOT_START =
+            new ServiceException.Code(offset(241), "receiver not start");
+    public static final ServiceException.Code RECEIVER_EXECUTION_FAILED =
+            new ServiceException.Code(offset(242), "receiver execution failed");
 
     private static int offset(int i) {
         return EXCEPTION_CODE_OFFSET + i;
@@ -127,6 +132,9 @@ public final class ServiceExceptionCodes {
         INVALID_JUDGEMENT_MODAL_UPDATE_HAPPENED_DATE.setCode(offset(210));
         INVALID_JUDGEMENT_MODAL_UPDATE_VALUE.setCode(offset(220));
         INVALID_ALARM_MODAL_UPDATE_HAPPENED_DATE.setCode(offset(230));
+        RECEIVER_FAILED.setCode(offset(240));
+        RECEIVER_NOT_START.setCode(offset(241));
+        RECEIVER_EXECUTION_FAILED.setCode(offset(242));
     }
 
     private ServiceExceptionCodes() {