+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/handler/ReceiveHandler.java
@@ -1,0 +1,12 @@
+package com.dwarfeng.judge.stack.handler;
+
+import com.dwarfeng.subgrade.stack.handler.StartableHandler;
+
+/**
+ * 接收处理器。
+ *
+ * @author DwArFeng
+ * @since 2.0.0
+ */
+public interface ReceiveHandler extends StartableHandler {
+}