+++ CHANGELOG.md
@@ -8,6 +8,12 @@
   - com.dwarfeng.judge.impl.service.telqos.TaskCheckCommand。
   - com.dwarfeng.judge.impl.service.telqos.JobLocalCacheCommand。
 
+- 实现预设接收器。
+  - com.dwarfeng.judge.impl.handler.receiver.DoNothingReceiver。
+  - com.dwarfeng.judge.impl.handler.receiver.DubboReceiver。
+  - com.dwarfeng.judge.impl.handler.receiver.InjvmReceiver。
+  - com.dwarfeng.judge.impl.handler.receiver.KafkaReceiver。
+
 - 实现预设判断器。
   - com.dwarfeng.judge.impl.handler.judger.groovy.GroovyJudgerRegistry。
 
@@ -15,6 +21,7 @@
   - com.dwarfeng.judge.impl.handler.analyser.groovy.GroovyAnalyserRegistry。
 
 - 实现核心机制。
+  - 接收机制。
   - 任务检查机制。
   - 作业机制。
   - 判断机制。